+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Solid Color Texture analyzer now works for all texture types (2D, 2DArray, 3D, Cube) in Unity 2019.2 and above
 
 ### Fixed
+* UI default column sorting
+* UI sorting criteria persistence after domain reload
 * Ignoring texture streaming issues
 * Texture size reporting on Unity 2022.2 and above is now accurate